refactor: we can essentially cast the ndarray to integers after scaling, so we don't have to coerce/cast later

+++ display.py
@@ -90,11 +90,11 @@ if __name__ == '__main__':
     elif which_color_map == 'uniform':
         # use evenly-spaced colors from the jet colormap
         # todo think about changing the linear space so it produces numbers in the 0..255 range
-        colors = 255.0 * cm.jet(np.linspace(0, 1, data_df['part_of_speech'].nunique()))
+        colors = (255.0 * cm.jet(np.linspace(0, 1, data_df['part_of_speech'].nunique()))).astype(int)
         part_of_speech_color_map = {
-            item[0]: '#{:02x}{:02x}{:02x}'.format(int(colors[index][0]),
-                                                  int(colors[index][1]),
-                                                  int(colors[index][2]))
+            item[0]: '#{:02x}{:02x}{:02x}'.format(colors[index][0],
+                                                  colors[index][1],
+                                                  colors[index][2])
             for index, item in enumerate(data_df['part_of_speech'].value_counts().items())}
     else:
         raise NotImplementedError('color map: {}'.format(which_color_map))
